context 적용
