+++ src/components/MonthNavigation.jsx
@@ -1,8 +1,12 @@
+import { useContext } from "react";
 import { MonthButtonContainer, MonthButton } from "../styles/HomeStyles";
+import { ExpenseContext } from "../utils/ExpenseContext";
 
 const MONTHS = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12];
 
-const MonthNavigation = ({ selectedMonth, setSelectedMonth }) => {
+const MonthNavigation = () => {
+  const { selectedMonth, setSelectedMonth } = useContext(ExpenseContext);
+
   return (
     <MonthButtonContainer>
       {MONTHS.map((element) => {
